TS-13 64bit support for x86_64 linux - tested on RHEL[4-5] 32-bit and FC11 32bit and 64bit with and without -m32 - patch sumbitted by John Plevyak - minor patches by Bryan Call, Leif Hedstrom, and Mladen Turk
git-svn-id: https://svn.apache.org/repos/asf/incubator/trafficserver/traffic/trunk@834757 13f79535-47bb-0310-9956-ffa450edef68

+++ iocore/cluster/P_ClusterCache.h
@@ -1117,7 +1117,7 @@ struct PingMessage:public ClusterMessageHeader
   {
     PingMessage *p = 0;
     // Maybe use offsetof here instead. /leif
-    return (int) (&p->data[0]);
+    return (uintptr_t) (&p->data[0]);
   }
   void init(inku16 vers = PING_MESSAGE_VERSION) {
     _init(vers);
